test : customer singUp test case 추가
